requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiosqlite==0.22.1
    # via llm-p (pyproject.toml)
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
bcrypt==4.3.0
    # via
    #   llm-p (pyproject.toml)
    #   passlib
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
cffi==2.0.0
    # via cryptography
click==8.3.2
    # via uvicorn
cryptography==46.0.7
    # via python-jose
dnspython==2.8.0
    # via email-validator
ecdsa==0.19.2
    # via python-jose
email-validator==2.3.0
    # via pydantic
fastapi==0.136.0
    # via llm-p (pyproject.toml)
greenlet==3.4.0
    # via
    #   llm-p (pyproject.toml)
    #   sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via llm-p (pyproject.toml)
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
passlib==1.7.4
    # via llm-p (pyproject.toml)
pyasn1==0.6.3
    # via
    #   python-jose
    #   rsa
pycparser==3.0
    # via cffi
pydantic==2.13.2
    # via
    #   llm-p (pyproject.toml)
    #   fastapi
    #   pydantic-settings
pydantic-core==2.46.2
    # via pydantic
pydantic-settings==2.13.1
    # via llm-p (pyproject.toml)
python-dotenv==1.2.2
    # via
    #   pydantic-settings
    #   uvicorn
python-jose==3.5.0
    # via llm-p (pyproject.toml)
python-multipart==0.0.26
    # via llm-p (pyproject.toml)
pyyaml==6.0.3
    # via uvicorn
rsa==4.9.1
    # via python-jose
ruff==0.15.11
    # via llm-p (pyproject.toml)
six==1.17.0
    # via ecdsa
sqlalchemy==2.0.49
    # via llm-p (pyproject.toml)
starlette==1.0.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
uvicorn==0.44.0
    # via llm-p (pyproject.toml)
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
